[Tech#429] Remove theme switch popup
