+++ spark/Demo/Classes/Tabbar/SparkTabbarController.swift
@@ -49,7 +49,6 @@ final class SparkTabbarController: UITabBarController {
 
         self.loadSparkConfiguration()
         self.setUpControllers()
-        self.addObservers()
         self.addPublishers()
     }
 
@@ -65,12 +64,6 @@ final class SparkTabbarController: UITabBarController {
         ]
     }
 
-    private func addObservers() {
-        NotificationCenter.default.addObserver(forName: UIDevice.deviceDidShakeNotification, object: nil, queue: .main) { _ in
-            self.presentThemeSwitchViewController()
-        }
-    }
-
     private func addPublishers() {
         self.themePublisher
             .$theme
@@ -79,21 +72,5 @@ final class SparkTabbarController: UITabBarController {
             self.tabBar.tintColor = theme.colors.main.main.uiColor
         }
         .store(in: &cancellables)
-    }
-}
-
-// MARK: - Navigation
-extension SparkTabbarController {
-
-    private func presentThemeSwitchViewController() {
-        let themeSwitchViewController = UIHostingController(
-            rootView: ThemeSwitchView(
-                dismissAction: { self.presentedViewController?.dismiss(animated: false) }
-            )
-        )
-        themeSwitchViewController.modalPresentationStyle = .overCurrentContext
-        themeSwitchViewController.modalTransitionStyle = .crossDissolve
-        themeSwitchViewController.view.backgroundColor = .clear
-        self.present(themeSwitchViewController, animated: true)
     }
 }
